requirements.txt
pandas
openpyxl
msoffcrypto-tool